chore: Update CSS files
This commit updates the CSS files in the client/src directory. It removes the overflow hidden property from the body element in index.css and adds comments to disable the scrollbar styling. It also updates the footer text in Footer.jsx to replace "BrandName" with "CodyX". Additionally, it adds an import statement for the Features component in HomePage.jsx and creates new CSS files for the About and Features components. Finally, it updates the dependencies in package.json to include "@reduxjs/toolkit" and "react-redux".